Move MergeSources to a new file and rename to IntersectGraphs.
